+++ src/interfaces/Cad.ts
@@ -19,6 +19,7 @@ export interface Cad {
   weight_prefix: string;
   height_prefix: string;
   assigned_status: string;
+  on_duty_status: string;
 
   version: {
     version: string;